Dev Update: Version setting, Thermostat fix, ...
Simplified bounded device print as the structure is common for any EP type
Allowed setting custom app version for EP, default is 0
Small fixes and code updates

+++ libraries/Zigbee/src/ep/ep_temperature_sensor.h
@@ -13,8 +13,8 @@ class ZigbeeTempSensor : public Zigbee_EP {
     // Set the temperature value in 0,01°C
     void setTemperature(float value);
 
-    // Set the min and max value for the temperature sensor in °C
-    void setMinMaxValue(int16_t min, int16_t max);
+    // Set the min and max value for the temperature sensor in 0,01°C
+    void setMinMaxValue(float min, float max);
 
     // Set the tolerance value for the temperature sensor in 0,01°C
     void setTolerance(float tolerance);
